Gru: change the sd power sequency
In the safety considerations, we should make sure the slot of SD is
enabled first, since we want to the power switch of corresponding is
powered up.

The different boards have the different power switch for sdmmc.
Some power switch IC need turn on delay for long time.

let's move the slot power of SD to romstage and avoid explicit delays
or per-board.

[Resolved the cherry-pick conflicts. The caused by TOT branch commit
7751c63 "UPSTREAM: lib: put romstage_handoff implementation in own
compilation unit", we needn't cherry-pick these patches for gru branch.]

BRANCH=none
BUG=b:35813418, b:35573103
TEST=check the signal for children of gru, and boot up from sd card.

Change-Id: I48ab543143d3de9be46608fc12d78e09decf8d79
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/447076
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/451020
Reviewed-by: Douglas Anderson <dianders@chromium.org>