UPSTREAM: ifdtool: avoid potential unaligned pointer usage
In get_region, ifdtool assigns a not-known-to-be-aligned
pointer to a uint32_t *. Now you know and I know that it is
almost certainly aligned, but clang on OSX doesn't like this,
and it's a dicey thing to do in any event, just waiting
to hit someone hard at some future date.

Assign the pointer to a void * and use memmove to copy
the value to a uint32_t.

This usage is more portable to all little-endian architectures,
now, but is still not endian-safe. I doubt we'll ever care.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4ae0bb5a44b085046a405e62c21fccbbae9828ca
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Original-Commit-Id: 33d4ebe3045c5f3bbc6f8edae0b76da85eacc388
Original-Change-Id: I85be85b9e78c74dde72584f848441a6ddd0ec0c2
Original-Signed-off-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Original-Commit-Id: 8db3c2a48549f28e40febae827666fe1730ebcd0
Original-Original-Change-Id: Ifb2f260c3363ab0f5b4a59e5a4e0b5ecf049fa96
Original-Original-Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Original-Original-Reviewed-on: https://review.coreboot.org/19921
Original-Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Original-Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/522709
Reviewed-on: https://chromium-review.googlesource.com/524355
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Duncan Laurie <dlaurie@google.com>
Tested-by: Duncan Laurie <dlaurie@google.com>
Trybot-Ready: Duncan Laurie <dlaurie@google.com>