mb/google/eve: Enable VMX
This feature was enabled at the kernel level, but that is triggering an
issue where FSP expects it to be disabled so it forces a cold reboot on
every warm reboot.  Since we want this enabled anyway just set it this
way in the BIOS so it matches what the kernel expects.

BUG=b:68666100
BRANCH=eve
TEST=pass firmware_FWtries on Eve with R63 OS image

Change-Id: I964d3d30392d130e808f37a661f2c89ec926cf58
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/749733
Reviewed-by: Furquan Shaikh <furquan@chromium.org>